backend/pyproject.toml
[tool.poetry]
name = "IO"
version = "1.0.0"
description = "Projekt na zaliczenie IO"
authors = ["Filip Ślęzak <filipslezakworks@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6,<4.0"
Django = "^3.2.4"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
pylint = "^2.8.3"
autopep8 = "^1.5.7"
rope = "^0.19.0"
